pyproject.toml
[project]
name = "olx-bot"
version = "0.1.0"
description = "Bot for parsing OLX search and send messages in telegram"
authors = [
    {name = "Bohdan Kokovych",email = "kokovych@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "aiogram (>=3.22.0,<4.0.0)",
    "python-dotenv (>=1.1.1,<2.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.group.dev.dependencies]
pytest = "^8.4.2"
black = "^25.1.0"
flake8 = "^7.3.0"
isort = "^6.0.1"
mypy = "^1.17.1"
pre-commit = "^4.3.0"

[tool.black]
line-length = 120
target-version = ["py312"]

[tool.isort]
profile = "black"

[tool.flake8]
max-line-length = 120
extend-ignore = ["E203", "W503"]

[tool.mypy]
python_version = "3.12"
warn_unused_configs = true
warn_return_any = true
disallow_untyped_defs = true

